Refactor RacingGame: replace MainWindow with RacingGame class, update UI and resource files, and remove unused MainWindow implementation
