+++ .vscode/settings.json
@@ -3,6 +3,8 @@
     "qmainwindow": "cpp",
     "qgraphicsrotation": "cpp",
     "*.tcc": "cpp",
-    "fstream": "cpp"
+    "fstream": "cpp",
+    "qtransform": "cpp",
+    "qgraphicsscene": "cpp"
   }
 }
