+++ src/main.cpp
@@ -1,11 +1,11 @@
-#include "../include/mainwindow.h"
+#include "../include/RacingGame.h"
 
 #include <QApplication>
 
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
-    MainWindow w;
+    RacingGame w;
 
     w.show();
