+++ ui/RacingGame.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>MainWindow</class>
- <widget class="QMainWindow" name="MainWindow">
+ <class>RacingGame</class>
+ <widget class="QMainWindow" name="RacingGame">
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -99,7 +99,6 @@
      </rect>
     </property>
     <property name="text">
-     <string>TEST</string>
      <string>Pause</string>
     </property>
    </widget>
